pyproject.toml
[tool.poetry]
name = "pyinterboleto"
version = "0.3.0"
description = "Biblioteca para facilitar o manuseio de boletos de contas PJ no Banco Inter."
authors = ["Felipe Souza Lima <feslima93@gmail.com>"]
license = "LGPL-3.0"
readme = "README.md"
repository = "https://github.com/feslima/pyinterboleto"
classifiers = ["Operating System :: OS Independent"]
packages = [{ include = "pyinterboleto", from = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
dacite = "^1.6.0"
tabulate = "^0.8.9"
Unidecode = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.4"
python-dotenv = "^0.19.0"
prettyprinter = "^0.18.0"
pre-commit = "^2.20.0"
autoflake = "^1.5.3"
isort = "^5.10.1"
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/feslima/pyinterboleto/issues"

[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_tests.py tests/*.py"
